memberService 에러 수정

+++ src/main/java/com/kosa/project/service/MemberService.java
@@ -5,9 +5,9 @@ import com.kosa.project.domain.MemberVO;
 public interface MemberService {
 	public Integer insert(MemberVO vo);
 
-	public MemberVO find(Long idx);
+	public MemberVO find(int idx);
 
-	public int remove(Long idx);
+	public int remove(int idx);
 
 	public int modify(MemberVO vo);
 }
